Refactoring, add interfaces, split into separate files
