+++ bin/no-skipped-tests.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+'use strict';
+
+require( './../dist/index.js' );
