+++ src/constants/default-options.ts
@@ -1,0 +1,8 @@
+import { NoSkippedTestsOptions } from './../interfaces/options.interface';
+
+/**
+ * Default options
+ */
+export const defaultOptions: NoSkippedTestsOptions = {
+	pattern: 'src/**/*.spec.ts' // TODO: Include js specs ...
+};
