+++ src/interfaces/no-skipped-tests-analyzer-error.interface.ts
@@ -1,0 +1,8 @@
+/**
+ * Error object as the result of the no skipped tests analsis
+ */
+export interface NoSkippedTestsAnalyzerError {
+	char: number;
+	identifier: string;
+	line: number;
+};
