+++ src/interfaces/no-skipped-tests-analyzer-result.interface.ts
@@ -1,0 +1,9 @@
+import { NoSkippedTestsAnalyzerError } from './no-skipped-tests-analyzer-error.interface';
+
+/**
+ * No skipped tests analyzer result object
+ */
+export interface NoSkippedTestsAnalyzerResult {
+	filePath: string;
+	errors: Array<NoSkippedTestsAnalyzerError>;
+}
