+++ src/interfaces/options.interface.ts
@@ -1,0 +1,6 @@
+/**
+ * No Skipped Tests Options
+ */
+export interface NoSkippedTestsOptions {
+	pattern?: string;
+}
